Rename kato/GoogleConfig and oort/GoogleConfig to KatoGoogleConfig and OortGoogleConfig to remove ambiguity from Spring bean resolution. Fixup yaml.
